requirements.txt
ctransformers
sentence-transformers==3.2.0
pinecone
langchain
langchain_community
pypdf
langchain-text-splitters==0.2.4
langchain-pinecone==0.1.3
streamlit
python-dotenv

-e .